remove caffe importer
