+++ modules/dnn/include/opencv2/dnn/all_layers.hpp
@@ -55,10 +55,9 @@ CV__DNN_INLINE_NS_BEGIN
 
   Classes listed here, in fact, provides C++ API for creating instances of built-in layers.
   In addition to this way of layers instantiation, there is a more common factory API (see @ref dnnLayerFactory), it allows to create layers dynamically (by name) and register new ones.
-  You can use both API, but factory API is less convenient for native C++ programming and basically designed for use inside importers (see @ref readNetFromCaffe(), @ref readNetFromTensorflow()).
+  You can use both API, but factory API is less convenient for native C++ programming and basically designed for use inside importers (see @ref readNetFromTensorflow()).
 
-  Built-in layers partially reproduce functionality of corresponding ONNX, TensorFlow and Caffe layers.
-  In particular, the following layers and Caffe importer were tested to reproduce <a href="http://caffe.berkeleyvision.org/tutorial/layers.html">Caffe</a> functionality:
+  Built-in layers partially reproduce functionality of corresponding ONNX, TensorFlow layers.
   - Convolution
   - Deconvolution
   - Pooling
