+++ modules/objdetect/src/barcode_decoder/common/super_scale.cpp
@@ -20,7 +20,7 @@ constexpr static float MAX_SCALE = 4.0f;
 
 int SuperScale::init(const std::string &proto_path, const std::string &model_path)
 {
-    srnet_ = dnn::readNetFromCaffe(proto_path, model_path);
+    srnet_ = dnn::readNet(proto_path, model_path);
     net_loaded_ = true;
     return 0;
 }
